Make two separate pages in search-bar-in-header example
