+++ examples/sandbox/src/pages/search-bar-in-header/search.js
@@ -1,0 +1,173 @@
+import React from "react";
+import { config } from "./config";
+
+import {
+  ErrorBoundary,
+  Facet,
+  SearchProvider,
+  SearchBox,
+  Results,
+  PagingInfo,
+  ResultsPerPage,
+  Paging,
+  Sorting,
+  WithSearch
+} from "@elastic/react-search-ui";
+import {
+  BooleanFacet,
+  Layout,
+  SingleSelectFacet,
+  SingleLinksFacet
+} from "@elastic/react-search-ui-views";
+import "@elastic/react-search-ui-views/lib/styles/styles.css";
+
+import Header from "./Header";
+
+const SORT_OPTIONS = [
+  {
+    name: "Relevance",
+    value: []
+  },
+  {
+    name: "Title",
+    value: [
+      {
+        field: "title",
+        direction: "asc"
+      }
+    ]
+  },
+  {
+    name: "State",
+    value: [
+      {
+        field: "states",
+        direction: "asc"
+      }
+    ]
+  },
+  {
+    name: "State -> Title",
+    value: [
+      {
+        field: "states",
+        direction: "asc"
+      },
+      {
+        field: "title",
+        direction: "asc"
+      }
+    ]
+  },
+  {
+    name: "Heritage Site -> State -> Title",
+    value: [
+      {
+        field: "world_heritage_site",
+        direction: "asc"
+      },
+      {
+        field: "states",
+        direction: "asc"
+      },
+      {
+        field: "title",
+        direction: "asc"
+      }
+    ]
+  }
+];
+
+export default function App() {
+  return (
+    <div>
+      <Header />
+      <SearchProvider config={config}>
+        <WithSearch mapContextToProps={({ wasSearched }) => ({ wasSearched })}>
+          {({ wasSearched }) => {
+            return (
+              <div className="App">
+                <ErrorBoundary>
+                  <Layout
+                    header={
+                      <SearchBox
+                        autocompleteMinimumCharacters={3}
+                        //searchAsYouType={true}
+                        autocompleteResults={{
+                          linkTarget: "_blank",
+                          sectionTitle: "Results",
+                          titleField: "title",
+                          urlField: "nps_link",
+                          shouldTrackClickThrough: true,
+                          clickThroughTags: ["test"]
+                        }}
+                        autocompleteSuggestions={true}
+                        debounceLength={0}
+                      />
+                    }
+                    sideContent={
+                      <div>
+                        {wasSearched && (
+                          <Sorting
+                            label={"Sort by"}
+                            sortOptions={SORT_OPTIONS}
+                          />
+                        )}
+                        <Facet
+                          field="states"
+                          label="States"
+                          filterType="any"
+                          isFilterable={true}
+                        />
+                        <Facet
+                          field="world_heritage_site"
+                          label="World Heritage Site?"
+                          view={BooleanFacet}
+                        />
+                        <Facet
+                          field="visitors"
+                          label="Visitors"
+                          view={SingleLinksFacet}
+                        />
+                        <Facet
+                          field="date_established"
+                          label="Date Established"
+                          filterType="any"
+                        />
+                        <Facet
+                          field="location"
+                          label="Distance"
+                          filterType="any"
+                        />
+                        <Facet
+                          field="acres"
+                          label="Acres"
+                          view={SingleSelectFacet}
+                        />
+                      </div>
+                    }
+                    bodyContent={
+                      <Results
+                        titleField="title"
+                        urlField="nps_link"
+                        thumbnailField="image_url"
+                        shouldTrackClickThrough={true}
+                      />
+                    }
+                    bodyHeader={
+                      <React.Fragment>
+                        {wasSearched && <PagingInfo />}
+                        {wasSearched && <ResultsPerPage />}
+                      </React.Fragment>
+                    }
+                    bodyFooter={<Paging />}
+                  />
+                </ErrorBoundary>
+              </div>
+            );
+          }}
+        </WithSearch>
+      </SearchProvider>
+    </div>
+  );
+}
